Create process governing susceptible to exposed disease state transition

Create process governing susceptible to exposed disease state transition

Usage

create_SE_process(variables_list, events_list, parameters_list, renderer)

Arguments

variables_list

A list of model variables as generated by create_variables()

events_list

A list of the model events as generated by create_events()

parameters_list

A list of model parameters as generated by get_parameters()

renderer

An individual-package render object

See also