Generate a vector of the initial disease states of all individuals in the population

Generate a vector of the initial disease states of all individuals in the population

Usage

generate_initial_disease_states(parameters_list)

Arguments

parameters_list

A list of model parameters as generated by get_parameters

See also