Establish the list of model parameters

This function creates a named list of model parameters which are to be used in the model. For example, the output of get_parameters() provides input to functions such as create_variables() and create_events().

get_parameters(overrides = list())

Arguments

overrides

A named list of parameters values to be used instead of the defaults. These parameters are:

human_population: the number of humans to include in the model

initial_proportion_child: proportion of population initially in the 'child' age class

initial_proportion_adult: proportion of population initially in the 'adult' age class

initial_proportion_elderly: proportion of population initially in the 'elderly' age class

number_initially_exposed: number of humans initially exposed (state = E)

seed: a seed to run the simulation with

mean_household_size: TBD

workplace_prop_max: maximum size of a workplace as a proportion of total adult population size

workplace_a: the a parameter for the Zipf-like distribution on workplace size

workplace_c: the c parameter for the Zipf-like distribution on workplace size

school_prop_max: maximum size of a school as a proporiton of total child population size

school_meanlog: the meanlog parameter for the log-normal distribution on school size

school_sdlog: the sdlog parameter for the log-normal distribution on school size

school_student_staff_ratio: the number of students to each adult staff member

leisure_mean_number_settings: TBD

leisure_mean_size: TBD

leisure_overdispersion_size: TBD

leisure_prop_max: TBD

duration_exposed: TBD

duration_infectious: TBD

beta_household: TBD

beta_workplace: TBD

beta_school: TBD

beta_leisure: TBD

beta_community: TBD

dt: TBD

simulation_time: TBD

Rendering Parameters

render_diagnostics: FALSE

Far UVC Intervention Parameters:

far_uvc_workplace: boolean switch set to TRUE if far UVC intervention parameterised in the workplace setting using set_ucv(); default = FALSE

far_uvc_workplace_coverage_type:Type of coverage. Select "random" for random selecting workplaces for UVC interventions and "targeted" for targeting interventions at most populous workplaces

far_uvc_workplace_coverage: Proportion of workplaces covered with far UVC (must be a numeric value between 0 and 1)

far_uvc_workplace_efficacy: : Efficacy of far UVC in the workplace setting (must be a numeric value between 0 and 1)

far_uvc_workplace_timestep: The timestep on which far UVC is implemented in the workplace setting (must be a numeric value greater than or equal to 0)

far_uvc_school: boolean switch set to TRUE if far UVC intervention parameterised in the school setting using set_ucv(); default = FALSE

far_uvc_school_coverage_type:Type of coverage. Select "random" for random selecting schools for UVC interventions and "targeted" for targeting interventions at most populous schools

far_uvc_school_coverage: Proportion of schools covered with far UVC (must be a numeric value between 0 and 1)

far_uvc_school_efficacy: : Efficacy of far UVC in the school setting (must be a numeric value between 0 and 1)

far_uvc_school_timestep: The timestep on which far UVC is implemented in the school setting (must be a numeric value greater than or equal to 0)

far_uvc_leisure: boolean switch set to TRUE if far UVC intervention parameterised in the leisure setting using set_ucv(); default = FALSE

far_uvc_leisure_coverage_type:Type of coverage. Select "random" for random selecting leisure settings for UVC interventions and "targeted" for targeting interventions at most populous leisure settings

far_uvc_leisure_coverage: Proportion of leisure settings covered with far UVC (must be a numeric value between 0 and 1)

far_uvc_leisure_efficacy: : Efficacy of far UVC in the leisure setting (must be a numeric value between 0 and 1)

far_uvc_leisure_timestep: The timestep on which far UVC is implemented in the leisure setting (must be a numeric value greater than or equal to 0)

far_uvc_household: boolean switch set to TRUE if far UVC intervention parameterised in the household setting using set_ucv(); default = FALSE

far_uvc_household_coverage_type:Type of coverage. Select "random" for random selecting households for UVC interventions and "targeted" for targeting interventions at most populous households

far_uvc_household_coverage: Proportion of households covered with far UVC (must be a numeric value between 0 and 1)

far_uvc_household_efficacy: : Efficacy of far UVC in the household setting (must be a numeric value between 0 and 1)

far_uvc_household_timestep: The timestep on which far UVC is implemented in the household setting (must be a numeric value greater than or equal to 0)