Sample variables from an log-normal distribution with fixed sum

The final sample is adjusted to ensure that the total sum of samples equals to N. All samples are rounded.

Usage

sample_log_normal(N, prop_max = 0.1, meanlog, sdlog)

Arguments

N

An integer giving the sum of generated random variables

prop_max

The proportion of N which is the maximum value of a single sample

meanlog

See the meanlog argument of dlnorm()

sdlog

See the sdlog argument of dlnorm()

See also