Update model parameters with far UVC settings

The set_uvc() function is a user-facing function that is used to parameterise far UVC deployment for an individual setting type (e.g. workplace). The function takes as arguments a helios parameter list, the setting type for which far UVC deployment is being parameterised, the coverage of far UVC within each location of a setting type (e.g. coverage of individual workplaces within the workplace setting class), the coverage_type (currently supporting random and targeted at the most populated settings), the efficacy of far UVC in the setting type, and the timestep on which far UVC is deployed. The function appends these additional setting-specific parameters to the parameter list and returns an updated version of it.

Usage

set_uvc(
  parameters_list,
  setting,
  coverage,
  coverage_target,
  coverage_type,
  efficacy,
  timestep
)

Arguments

parameters_list

A list of parameters as generated using get_parameters()

setting

A character string describing the setting type in which far UVC is being deployed

coverage

A numeric value describing the coverage of far UVC within the setting class for which which far UVC is deployed

coverage_target

A character describing the target of the coverage ("buildings" or "individuals")

coverage_type

A character describing the type of coverage ("random" or "targeted")

efficacy

A numeric value describing the efficacy of the far UVC deployed

timestep

A numeric value describing the timestep in which far UVC is deployed

See also