Establish the list of model parameters

This function creates a named list of model parameters which are to be used in the model. For example, the output of get_parameters() provides input to functions such as create_variables() and create_events().

Usage

get_parameters(overrides = list(), archetype = "none")

Arguments

overrides

A named list of parameters values to be used instead of the defaults. These parameters are:

human_population: the number of humans to include in the model

initial_proportion_child: proportion of population initially in the 'child' age class

initial_proportion_adult: proportion of population initially in the 'adult' age class

initial_proportion_elderly: proportion of population initially in the 'elderly' age class

number_initial_S: number of humans initially Susceptible (state = S)

number_initial_E: number of humans initially Exposed (state = E)

number_initial_I: number of humans initially Infectious (state = I)

number_initial_R: number of humans initially Recovered (state = R)

seed: a seed to run the simulation with

mean_household_size: TBD

workplace_prop_max: maximum size of a workplace as a proportion of total adult population size

workplace_a: the a parameter for the Zipf-like distribution on workplace size

workplace_c: the c parameter for the Zipf-like distribution on workplace size

school_prop_max: maximum size of a school as a proporiton of total child population size

school_meanlog: the meanlog parameter for the log-normal distribution on school size

school_sdlog: the sdlog parameter for the log-normal distribution on school size

school_student_staff_ratio: the number of students to each adult staff member

leisure_mean_number_settings: TBD

leisure_mean_size: TBD

leisure_overdispersion_size: TBD

leisure_prop_max: TBD

duration_exposed: TBD

duration_infectious: TBD

prob_inf_external: The probability a susceptible individual is infected from an external source

beta_household: TBD

beta_workplace: TBD

beta_school: TBD

beta_leisure: TBD

beta_community: TBD

dt: TBD

simulation_time: TBD

household_distribution_country: TBD

school_distribution_country: TBD

workplace_distribution_country: TBD

endemic_or_epidemic: TBD

duration_immune: TBD

prob_inf_external: TBD

Rendering Parameters

render_diagnostics: FALSE

Far UVC Intervention Parameters:

far_uvc_workplace: boolean switch set to TRUE if far UVC intervention parameterised in the workplace setting using set_ucv(); default = FALSE

far_uvc_workplace_coverage: Proportion of coverage of far UVC (must be a numeric value between 0 and 1)

far_uvc_workplace_coverage_target: The target that coverage proportion applies to. Select either "individuals" or "buildings"

far_uvc_workplace_coverage_type: Type of coverage. Select "random" for random selecting workplaces for UVC interventions and "targeted" for targeting interventions at most populous workplaces

far_uvc_workplace_efficacy: : Efficacy of far UVC in the workplace setting (must be a numeric value between 0 and 1)

far_uvc_workplace_timestep: The timestep on which far UVC is implemented in the workplace setting (must be a numeric value greater than or equal to 0)

far_uvc_school: boolean switch set to TRUE if far UVC intervention parameterised in the school setting using set_ucv(); default = FALSE

far_uvc_school_coverage: Proportion of schools covered with far UVC (must be a numeric value between 0 and 1)

far_uvc_school_coverage_target: The target that coverage proportion applies to. Select either "individuals" or "buildings"

far_uvc_school_coverage_type: Type of coverage. Select "random" for random selecting schools for UVC interventions and "targeted" for targeting interventions at most populous schools

far_uvc_school_efficacy: : Efficacy of far UVC in the school setting (must be a numeric value between 0 and 1)

far_uvc_school_timestep: The timestep on which far UVC is implemented in the school setting (must be a numeric value greater than or equal to 0)

far_uvc_leisure: boolean switch set to TRUE if far UVC intervention parameterised in the leisure setting using set_ucv(); default = FALSE

far_uvc_leisure_coverage: Proportion of leisure settings covered with far UVC (must be a numeric value between 0 and 1)

far_uvc_leisure_coverage_target: The target that coverage proportion applies to. Select either "individuals" or "buildings"

far_uvc_leisure_coverage_type: Type of coverage. Select "random" for random selecting leisure settings for UVC interventions and "targeted" for targeting interventions at most populous leisure settings

far_uvc_leisure_efficacy: : Efficacy of far UVC in the leisure setting (must be a numeric value between 0 and 1)

far_uvc_leisure_timestep: The timestep on which far UVC is implemented in the leisure setting (must be a numeric value greater than or equal to 0)

far_uvc_household: boolean switch set to TRUE if far UVC intervention parameterised in the household setting using set_ucv(); default = FALSE

far_uvc_household_coverage: Proportion of households covered with far UVC (must be a numeric value between 0 and 1)

far_uvc_household_coverage_target: The target that coverage proportion applies to. Select either "individuals" or "buildings"

far_uvc_household_coverage_type: Type of coverage. Select "random" for random selecting households for UVC interventions and "targeted" for targeting interventions at most populous households

far_uvc_household_efficacy: : Efficacy of far UVC in the household setting (must be a numeric value between 0 and 1)

far_uvc_household_timestep: The timestep on which far UVC is implemented in the household setting (must be a numeric value greater than or equal to 0)

Setting-Specific Riskiness Parameters:

setting_specific_riskiness_workplace: boolean switch set to TRUE if setting-specific riskiness parameterised in the workplace setting using set_setting_specific_riskiness(); default = FALSE

setting_specific_riskiness_workplace_meanlog: The mean of the truncated log-normal distribution from which the setting-specific riskiness of individual workplaces is drawn; default = 0

setting_specific_riskiness_workplace_sdlog: The standard deviation of the truncated log-normal distribution from which the setting-specific riskiness of individual workplaces is drawn; default = 0.37

setting_specific_riskiness_workplace_min: The minimum value for truncation on the left tail of the truncated log-normal distribution from which the setting-specific riskiness of individual workplaces is drawn; default = 0.4472

setting_specific_riskiness_workplace_max: The maximum value for truncation on the right tail of the truncated log-normal distribution from which the setting-specific riskiness of individual workplaces is drawn; default = 2.236

setting_specific_riskiness_school: boolean switch set to TRUE if setting-specific riskiness parameterised in the school setting using set_setting_specific_riskiness(); default = FALSE

setting_specific_riskiness_school_meanlog: The mean of the truncated log-normal distribution from which the setting-specific riskiness of individual schools is drawn; default = 0

setting_specific_riskiness_school_sdlog: The standard deviation of the truncated log-normal distribution from which the setting-specific riskiness of individual schools is drawn; default = 0.37

setting_specific_riskiness_school_min: The minimum value for truncation on the left tail of the truncated log-normal distribution from which the setting-specific riskiness of individual schools is drawn; default = 0.4472

setting_specific_riskiness_school_max: The maximum value for truncation on the right tail of the truncated log-normal distribution from which the setting-specific riskiness of individual schools is drawn; default = 2.236

setting_specific_riskiness_leisure: boolean switch set to TRUE if setting-specific riskiness parameterised in the leisure setting using set_setting_specific_riskiness(); default = FALSE

setting_specific_riskiness_leisure_meanlog: The mean of the truncated log-normal distribution from which the setting-specific riskiness of individual leisure locations is drawn; default = 0

setting_specific_riskiness_leisure_sdlog: The standard deviation of the truncated log-normal distribution from which the setting-specific riskiness of individual leisure settings is drawn; default = 0.37

setting_specific_riskiness_leisure_min: The minimum value for truncation on the left tail of the truncated log-normal distribution from which the setting-specific riskiness of individual leisure settings is drawn; default = 0.4472

setting_specific_riskiness_leisure_max:The maximum value for truncation on the right tail of the truncated log-normal distribution from which the setting-specific riskiness of individual leisure settings is drawn; default = 2.236

setting_specific_riskiness_household: boolean switch set to TRUE if setting-specific riskiness parameterised in the household setting using set_setting_specific_riskiness(); default = FALSE

setting_specific_riskiness_household_meanlog: The mean of the truncated log-normal distribution from which the setting-specific riskiness of individual households is drawn; default = 0

setting_specific_riskiness_household_sdlog: The standard deviation of the truncated log-normal distribution from which the setting-specific riskiness of individual households is drawn; default = 0.37

setting_specific_riskiness_household_min: The minimum value for truncation on the left tail of the truncated log-normal distribution from which the setting-specific riskiness of individual households is drawn; default = 0.4472

setting_specific_riskiness_household_max: The maximum value for truncation on the right tail of the truncated log-normal distribution from which the setting-specific riskiness of individual households is drawn; default = 2.236

archetype

A text string indicating the pathogen archetype parameter set to load (default = "none", current options are flu, sars_cov_2, and measles)